теперь точно хз

+++ temperature.py
@@ -1,4 +1,4 @@
 def convert_temperature(a):
-    return int(a) * 9 / 5 + 32
+    return float(a) * 9 / 5 + 32
 user_input = input("Введите температуру в градусах по Цельсию ")
 print(f"Это {convert_temperature(user_input)} градусов по Фаренгейту")
